SmartHomeIoT/raspberry-pi/requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
RPi.GPIO>=0.7.0; platform_machine == "armv7l" or platform_machine == "aarch64"
picamera2>=0.3.0; platform_machine == "armv7l" or platform_machine == "aarch64"
opencv-python>=4.5.0
Pillow>=9.0.0
